String extension attributedString 접근제어 변경, ExpositionUniverselleItem 파라미터타입 변경
