+++ Expo1900/Expo1900/Views/ExpositionTableViewCell.swift
@@ -12,15 +12,7 @@ final class ExpositionTableViewCell: UITableViewCell {
     @IBOutlet private weak var itemTitleLabel: UILabel!
     @IBOutlet private weak var itemSubtitleLabel: UILabel!
     
-    func configure(item: ExpositionUniverselleItem?) {
-        guard let item = item else {
-            itemImageView.image = nil
-            itemTitleLabel.text = nil
-            itemSubtitleLabel.text = nil
-            
-            return
-        }
-        
+    func configure(item: ExpositionUniverselleItem) {
         itemImageView.image = UIImage(named: item.imageName)
         itemTitleLabel.text = item.name
         itemSubtitleLabel.text = item.shortDescription
@@ -28,6 +20,9 @@ final class ExpositionTableViewCell: UITableViewCell {
     
     override func prepareForReuse() {
         super.prepareForReuse()
-        configure(item: nil)
+        
+        itemImageView.image = nil
+        itemTitleLabel.text = nil
+        itemSubtitleLabel.text = nil
     }
 }
